Refactor comment and tweet controllers

+++ src/controllers/tweet.controller.js
@@ -47,37 +47,62 @@ const getUserTweets = asyncHandler(async (req, res) => {
 const updateTweet = asyncHandler(async (req, res) => {
   //TODO: update tweet
   const { tweetId } = req.params
+  const { _id } = req.user
 
   if (!tweetId) {
     throw new ApiError(400, "Tweet ID is required")
   }
 
-  const tweet = await Tweet.findByIdAndUpdate(tweetId, {
+  const tweet = await Tweet.findById(tweetId);
+
+  if (!tweet) {
+    throw new ApiError(404, "Tweet not found")
+  }
+
+  if (tweet.owner.toString() !== _id.toString()) {
+    throw new ApiError(401, "You are not authorized to update this tweet")
+  }
+
+  const updateTweet = await Tweet.findByIdAndUpdate(tweetId, {
     $set: {
       content: req.body.content
     }
   }, {
     new: true
   })
-  if (!tweet) {
+  if (!updateTweet) {
     throw new ApiError(404, "Tweet not found")
   }
 
-  res.status(200).json(new ApiResponse(200, { tweet }, "Tweet updated successfully"))
+  res.status(200).json(new ApiResponse(200, {}, "Tweet updated successfully"))
 })
 
 const deleteTweet = asyncHandler(async (req, res) => {
   //TODO: delete tweet
-
   const { tweetId } = req.params
 
   if (!tweetId) {
     throw new ApiError(400, "Tweet ID is required")
   }
 
-  const tweet = await Tweet.findByIdAndDelete(tweetId)
+
+  if (!tweetId) {
+    throw new ApiError(400, "Tweet ID is required")
+  }
+
+  const tweet = await Tweet.findById(tweetId);
 
   if (!tweet) {
+    throw new ApiError(404, "Tweet not found")
+  }
+
+  if (tweet.owner.toString() !== _id.toString()) {
+    throw new ApiError(401, "You are not authorized to update this tweet")
+  }
+
+  const deleteTweet = await Tweet.findByIdAndDelete(tweetId)
+
+  if (!deleteTweet) {
     throw new ApiError(404, "Tweet not found")
   }
